Fixed some bugs

+++ Views/Room/Room.py
@@ -10,4 +10,4 @@ class RoomV:
 
     def render_tile_map(self, surface: pg.Surface) -> None:
         for loc in self.tile_map.map:
-            self.tile_map.map[loc].view.draw(surface, tuple(map(float, loc.split(';'))))
+            self.tile_map.map[loc].view.draw(surface, tuple(map(int, loc.split(';'))))
